% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrelsa.R
\name{hrelsa}
\alias{hrelsa}
\title{Relative Severity Assessment Score for Humans
= hRELSA}
\usage{
hrelsa(
  pre = pre,
  bsl,
  a = 1,
  drop = NULL,
  turnvars = NULL,
  ambivars = NULL,
  zvars = NULL,
  relsaNA = NA
)
}
\arguments{
\item{pre}{normalized data set}

\item{bsl}{data set baseline}

\item{a}{unique patient index in data set}

\item{drop}{variables which shall not used for hRELSA}

\item{turnvars}{variables with "turned" direction}

\item{ambivars}{ambivalent variables}

\item{zvars}{z variables}

\item{relsaNA}{code how to handle NaN values during calculations
(default is NA)}
}
\value{
\code{hrelsa} list with hRELSA results for specific patient
(differences and weights plus hRELSA score)
}
\description{
The \code{hrelsa} function calculates a composite relative severity score
based on normalized differences. Further, the values are regularized with a
reference set to estimate relative severity.
This is the main function of hRELSA and is normally used in
the hrelsa_final function.
}
