% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrelsa_analysis.R
\name{hrelsa_analysis}
\alias{hrelsa_analysis}
\title{Evaluates a analysis of the hRELSAed data set}
\usage{
hrelsa_analysis(set)
}
\arguments{
\item{final}{hRELSAed data set (by hrelsa_final)}
}
\value{
\code{analysis} Outputs some basic data set informations, and the 
analysis of wfactor, wf and rms
}
\description{
The \code{hrelsa_analysis} evaluates relevant informations about the patients
with the highest severity
}
