% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrelsa_baselines.R
\name{hrelsa_baselines}
\alias{hrelsa_baselines}
\title{Defines baselines and maximum severity values for hRELSA calculation}
\usage{
hrelsa_baselines(
  pre,
  bslday = NULL,
  vars = NULL,
  zvars = NULL,
  turnvars = NULL,
  ambivars = NULL
)
}
\arguments{
\item{pre}{normalized data set}

\item{bslday}{time point in the data set used for baseline calculation
(i.e. day=-1) (not needed with just z variables)}

\item{vars}{variables to use as references}

\item{zvars}{z variables (SDS variables)}

\item{turnvars}{variables with "turned" direction
(e.g. c("hr","temp")) (not needed with just z variables)}

\item{ambivars}{ambivalent variables}
}
\value{
\code{ret} list with reference variables set to 100 %,
maximum reached values and some characteristics
}
\description{
The \code{hrelsa_baselines} function calculates baselines, maximum severity,
maximum delta and some characteristics
}
