% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrelsa_days.R
\name{hrelsa_days}
\alias{hrelsa_days}
\title{Forms dates to days for hRELSA calculation}
\usage{
hrelsa_days(
  raw,
  format = "day",
  date_format = "\%d\%b\%Y",
  formthis = NULL,
  newdayone = TRUE
)
}
\arguments{
\item{raw}{data set}

\item{format}{either "day", "year" or "timecode" (a transformation to years is possible as well)}

\item{formthis}{date variable}

\item{newdayone}{for each patient a day 1? (TRUE/FALSE)}
}
\value{
\code{raw} data set with day column added
}
\description{
The \code{hrelsa_days} function transforms dates into days as integer
either with the first date of each patient as day 1
or with the first date of the whole data set as day 1
a day column will be added
}
