% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrelsa_final.R
\name{hrelsa_final}
\alias{hrelsa_final}
\title{Relative Severity Assessment Score for Humans
= hRELSA}
\usage{
hrelsa_final(
  pre,
  bsl,
  drop = NULL,
  turnvars = turnvars,
  zvars = zvars,
  relsaNA = NA
)
}
\arguments{
\item{pre}{normalized data set}

\item{bsl}{data set baseline}

\item{drop}{variables which shall not used for hRELSA}

\item{turnvars}{variables with "turned" direction}

\item{zvars}{z variables}

\item{relsaNA}{code how to handle NaN values during calculations
(default is NA)}
}
\value{
\code{final} list with hRELSA results for all patients
(differences and weights plus hRELSA score)
}
\description{
The \code{hrelsa_final} function calculates a composite relative severity score
based on normalized differences. Further, the values are regularized with a
reference set to estimate relative severity.
This is just the function to perform hRELSA on all patients.
The main function is hRELSA.
}
