% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrelsa_format.R
\name{hrelsa_format}
\alias{hrelsa_format}
\title{Formats the data set for hRELSA calculation}
\usage{
hrelsa_format(
  dat,
  id = NULL,
  treatment = NULL,
  condition = NULL,
  day = NULL,
  vars = NULL
)
}
\arguments{
\item{dat}{data set}

\item{id}{id variable}

\item{treatment}{treatment variable}

\item{condition}{condition variable}

\item{day}{day variable}

\item{vars}{variables to work with}
}
\value{
\code{dat} data set rearranged,
just with needed and with properly named columns
}
\description{
The \code{hrelsa_format} renames and rearranges the columns to
proper work with
}
