% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrelsa_norm.R
\name{hrelsa_norm}
\alias{hrelsa_norm}
\title{Normalizes the data set for hRELSA calculation}
\usage{
hrelsa_norm(set, normthese = NULL, zvars = NULL, ontime = NULL)
}
\arguments{
\item{set}{data set}

\item{normthese}{variable names in the data set that shall be normalized}

\item{zvars}{z variables (need to be contained in normthese as well)}

\item{ontime}{unique reference time point in data set (e.g. 1 for day 1)}
}
\value{
\code{set} data set with normalized values
}
\description{
The \code{hrelsa_norm} function normalizes defined variables to 100 %
at a given time point
when z variables are involved they have a special way of being normalized
z_norm = (|z| + 1) * 100
when normalizing z variables the reference time point does not matter
baseline for z variables is 0 (0 <- 100 %)
}
