% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrelsa_plot.R
\name{hrelsa_plot}
\alias{hrelsa_plot}
\title{Plot hRELSA score and variables together}
\usage{
hrelsa_plot(
  set,
  hRELSA,
  levels = NULL,
  a = 1,
  plotvar = NULL,
  plothRELSA = TRUE,
  myylim = c(-4, 4),
  myYlim = c(0, 2),
  mypch = 1,
  mycol = "red",
  myXlab = "time point"
)
}
\arguments{
\item{set}{data set}

\item{levels}{k+1 levels from kmeans calculations (hrelsa_levels function)}

\item{a}{which unique patient in the set?}

\item{plotvar}{which variable shall be plotted (does not scale with hRELSA score)}

\item{plothRELSA}{plot the hRELSA score?}

\item{myylim}{custom y-axis limits (left axis)}

\item{myYlim}{custom y-axis limits (right axis)}

\item{mypch}{custom dot marker}

\item{mycol}{custom plot color}

\item{myXlab}{custom x-axis label}

\item{RELSA}{hRELSA object}
}
\value{
\code{plotset} data.frame with not normalized values of unique patient
}
\description{
The \code{hrelsa_plot} function visualizes the hRELSA score alone or in
combination with one of the variables.
}
