% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrelsa_wrapper.R
\name{hrelsa_wrapper}
\alias{hrelsa_wrapper}
\title{The hRELSA wrapper function is currently OUT OF USE.
Please ask the author for application of the hRELSA.}
\usage{
hrelsa_wrapper(
  data_path = NULL,
  data_seperation = NULL,
  data_decimal = NULL,
  data_sheet = NULL,
  raw = NULL,
  id = NULL,
  treatment = NULL,
  condition = NULL,
  day = NULL,
  form_to_day = NULL,
  day_format = NULL,
  form_day_into = "day",
  new_day_one = NULL,
  vars = NULL,
  turnvars = NULL,
  zvars = NULL,
  dropvars = NULL,
  baseline_day = 1,
  levels = FALSE,
  k = NULL,
  plot_instead_of_scree = TRUE,
  which_patient = NULL,
  which_var_to_plot = NULL
)
}
\description{
The hRELSA wrapper function is currently OUT OF USE.
Please ask the author for application of the hRELSA.
}
