% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_adaptive_baselines_range.R
\name{pvs_adaptive_baselines_range}
\alias{pvs_adaptive_baselines_range}
\title{PVS}
\usage{
pvs_adaptive_baselines_range(
  dat,
  reference_dat = NULL,
  max_sev = NULL,
  vars = NULL,
  turnvars = NULL,
  ambivars = NULL,
  realtime = NULL,
  sex_dat = NULL,
  sex_data_id_col = NULL,
  sex_data_sex_col = NULL,
  norm_dat = NULL
)
}
\arguments{
\item{dat}{data set}

\item{reference_dat}{reference data set}

\item{max_sev}{if no reference data set is given, set max severity percentage}

\item{vars}{variables}

\item{turnvars}{turned variables}

\item{ambivars}{ambivalent variables}

\item{sex_dat}{data with date of birth in it}

\item{sex_data_id_col}{column id of id in sex data set}

\item{sex_data_sex_col}{column id of sex in sex data set}

\item{norm_dat}{data set with normal / reference values}

\item{norm_dat_names}{names of the norm dat data set}
}
\value{
\code{ret} list with reference variables set to 100 %,
maximum reached values and some characteristics
}
\description{
The \code{pvs_adaptive_baselines_sex} function normalizes the data,
calculates baselines, maximum severity, maximum delta and some characteristics
for data using the adaptive reference method
}
