% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_analysis.R
\name{pvs_analysis}
\alias{pvs_analysis}
\title{PVS}
\usage{
pvs_analysis(set)
}
\arguments{
\item{set}{data set with PVS values (by pvs_final)}
}
\value{
\code{analysis} basic data set information, and the
analysis of wfactor, wf and rms
}
\description{
The \code{pvs_analysis} evaluates relevant information about the patients
with the highest severity
}
