% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_baselines.R
\name{pvs_baselines}
\alias{pvs_baselines}
\title{PVS}
\usage{
pvs_baselines(
  pre,
  bslday = NULL,
  vars = NULL,
  zvars = NULL,
  turnvars = NULL,
  ambivars = NULL
)
}
\arguments{
\item{pre}{normalized data set}

\item{bslday}{time point in the data set used for baseline calculation}

\item{vars}{variables to use as references}

\item{zvars}{z variables}

\item{turnvars}{variables with "turned" direction}

\item{ambivars}{ambivalent variables}
}
\value{
\code{ret} list with reference variables set to 100 %,
maximum reached values and some characteristics
}
\description{
The \code{pvs_baselines} function calculates baselines, maximum severity,
maximum delta and some characteristics
}
