% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_final.R
\name{pvs_final}
\alias{pvs_final}
\title{PVS}
\usage{
pvs_final(
  pre,
  bsl,
  drop = NULL,
  turnvars = NULL,
  ambivars = NULL,
  zvars = NULL,
  pvsNA = NA
)
}
\arguments{
\item{pre}{normalized data set}

\item{bsl}{baseline}

\item{drop}{variables which shall not used for PVS}

\item{turnvars}{variables with "turned" direction}

\item{ambivars}{ambivalent variables}

\item{zvars}{z variables}

\item{pvsNA}{code how to handle NaN values during calculations
(default is NA)}
}
\value{
\code{final} list with PVS results for all patients
(differences and weights plus PVS score)
}
\description{
The \code{pvs_final} applies the PVS code on the whole data set
}
