% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_format.R
\name{pvs_format}
\alias{pvs_format}
\title{PVS}
\usage{
pvs_format(
  dat,
  id = NULL,
  treatment = NULL,
  condition = NULL,
  time = NULL,
  vars = NULL,
  included_realtime = NULL
)
}
\arguments{
\item{dat}{data set}

\item{id}{id variable}

\item{treatment}{treatment variable}

\item{condition}{condition variable}

\item{time}{time variable}

\item{vars}{variables}

\item{included_realtime}{variable of real time variable when needed}
}
\value{
\code{dat} formatted data set
}
\description{
The \code{pvs_format} formats the data
}
