% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvs_norm.R
\name{pvs_norm}
\alias{pvs_norm}
\title{PVS}
\usage{
pvs_norm(set, normthese = NULL, zvars = NULL, ambivars = NULL, ontime = NULL)
}
\arguments{
\item{set}{data set}

\item{normthese}{variables in the data set that shall be normalized}

\item{zvars}{z variables (need to be contained in normthese as well)}

\item{ambivars}{ambivalent variables}

\item{ontime}{unique reference time point in data set (e.g. 1 for day 1)}
}
\value{
\code{set} data set with normalized values
}
\description{
The \code{pvs_norm} function normalizes defined variables to 100 %
at a given time point
}
\details{
when z variables are involved they have a special way of being normalized
z_norm = (|z| + 1) * 100
when normalizing z variables the reference time point does not matter
baseline for z variables is 0 (0 <- 100 %)
}
