% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrelsa_adaptive_baselines.R
\name{hrelsa_adaptive_baselines}
\alias{hrelsa_adaptive_baselines}
\title{Defines baselines and maximum severity values for hRELSA calculation using
the adaptive reference method
Here is the normalization included as well}
\usage{
hrelsa_adaptive_baselines(
  dat,
  reference_dat,
  vars = NULL,
  turnvars = NULL,
  ambivars = NULL,
  realtime = NULL,
  dob_dat = NULL,
  dob_data_id_col = NULL,
  dob_data_dob_col = NULL,
  norm_dat = NULL,
  norm_dat_names = NULL
)
}
\arguments{
\item{dat}{formatted data set}

\item{vars}{variables to work with}

\item{turnvars}{turned variables}

\item{ambivars}{ambivalent variables}

\item{realtime}{the real time point for age calculcation, the real time needs to be included in dat}

\item{dob_dat}{data with date of birth in it}

\item{dob_data_id_col}{column id of id in dob data set}

\item{dob_data_dob_col}{column id of date of birth in dob data set}

\item{norm_dat}{data set with normal / reference values}

\item{norm_dat_names}{names of the norm dat data set}
}
\value{
\code{ret} list with reference variables set to 100 %,
maximum reached values and some characteristics
}
\description{
The \code{hrelsa_adaptive_baselines} function normalizes the data,
calculates baselines, maximum severity, maximum delta and some characteristics
}
