% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrelsa_adaptive_baselines.R
\name{hrelsa_adaptive_baselines}
\alias{hrelsa_adaptive_baselines}
\title{X}
\usage{
hrelsa_adaptive_baselines(
  dat,
  vars = NULL,
  turnvars = NULL,
  realtime = NULL,
  dob_dat = NULL,
  dob_data_id_col = NULL,
  dob_data_dob_col = NULL,
  norm_dat = NULL,
  norm_dat_names = NULL
)
}
\arguments{
\item{raw}{X}
}
\value{
\code{X} X
}
\description{
The \code{hrelsa_adaptive_baselines} X
}
